add logic to first time check
