+++ BoBNolepAssique/Controllers/Intro/DanaImpulsifViewController.swift
@@ -1,0 +1,17 @@
+//
+//  DanaImpulsifViewController.swift
+//  BoBNolepAssique
+//
+//  Created by Raja Azian on 17/06/21.
+//
+
+import UIKit
+
+class DanaImpulsifViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Dana Impulsif"
+    }
+
+}
