+++ BoBNolepAssique/Controllers/Intro/InvestasiViewController.swift
@@ -1,0 +1,17 @@
+//
+//  InvestasiViewController.swift
+//  BoBNolepAssique
+//
+//  Created by Raja Azian on 17/06/21.
+//
+
+import UIKit
+
+class InvestasiViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Investasi"
+    }
+
+}
