+++ BoBNolepAssique/Controllers/Intro/KebutuhanViewController.swift
@@ -1,0 +1,17 @@
+//
+//  KebutuhanViewController.swift
+//  BoBNolepAssique
+//
+//  Created by Raja Azian on 17/06/21.
+//
+
+import UIKit
+
+class KebutuhanViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Kebutuhan"
+    }
+
+}
