+++ BoBNolepAssique/Controllers/Intro/PendapatanViewController.swift
@@ -1,0 +1,18 @@
+//
+//  PendapatanViewController.swift
+//  BoBNolepAssique
+//
+//  Created by Raja Azian on 17/06/21.
+//
+
+import UIKit
+
+class PendapatanViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Pendapatan"
+    }
+    
+
+}
