+++ BoBNolepAssique/Controllers/Intro/PerkiraanBiayaViewController.swift
@@ -1,0 +1,17 @@
+//
+//  PerkiraanBiayaViewController.swift
+//  BoBNolepAssique
+//
+//  Created by Raja Azian on 17/06/21.
+//
+
+import UIKit
+
+class PerkiraanBiayaViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Perkiraan Biaya"
+    }
+
+}
