+++ BoBNolepAssique/SceneDelegate.swift
@@ -16,6 +16,20 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
+        
+        if isFirstOpenApp() {
+            let pendapatanStoryboard: UIStoryboard = UIStoryboard(name: "Pendapatan", bundle: nil)
+            guard let rootVC = pendapatanStoryboard.instantiateViewController(identifier: "PendapatanController") as? PendapatanViewController else {
+                print("ViewController not found")
+                return
+            }
+            let rootNC = UINavigationController(rootViewController: rootVC)
+            rootNC.navigationBar.prefersLargeTitles = true
+            self.window?.rootViewController = rootNC
+            self.window?.makeKeyAndVisible()
+        }
+        
+        
         guard let _ = (scene as? UIWindowScene) else { return }
     }
