+++ BoBNolepAssique/Views/Intro/Pendapatan.storyboard
@@ -11,7 +11,7 @@
         <!--Pendapatan View Controller-->
         <scene sceneID="s0d-6b-0kx">
             <objects>
-                <viewController id="Y6W-OH-hqX" customClass="PendapatanViewController" customModule="BoBNolepAssique" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="PendapatanController" id="Y6W-OH-hqX" customClass="PendapatanViewController" customModule="BoBNolepAssique" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="5EZ-qb-Rvc">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
